+++ src/fidesops/ops/tasks/__init__.py
@@ -1,12 +1,29 @@
-from typing import Any, Dict, MutableMapping
+from typing import Any, ContextManager, Dict, MutableMapping
 
-from celery import Celery
+from celery import Celery, Task
 from celery.utils.log import get_task_logger
 from fideslib.core.config import load_toml
+from fideslib.db.session import get_db_session
+from sqlalchemy.orm import Session
 
 from fidesops.ops.core.config import config
 
 logger = get_task_logger(__name__)
+
+EMAIL_QUEUE_NAME = "fidesops.email"
+
+
+class DatabaseTask(Task):  # pylint: disable=W0223
+    _session = None
+
+    @property
+    def session(self) -> ContextManager[Session]:
+        """Creates Session once per process"""
+        if self._session is None:
+            SessionLocal = get_db_session(config)
+            self._session = SessionLocal()
+
+        return self._session
 
 
 def _create_celery(config_path: str = config.execution.celery_config_path) -> Celery:
@@ -20,6 +37,8 @@ def _create_celery(config_path: str = config.execution.celery_config_path) -> Ce
         # Defaults for the celery config
         "broker_url": config.redis.connection_url,
         "result_backend": config.redis.connection_url,
+        # Fidesops requires this to route emails to separate queues
+        "task_create_missing_queues": True,
     }
 
     try:
@@ -47,7 +66,15 @@ celery_app = _create_celery()
 
 def start_worker() -> None:
     logger.info("Running Celery worker...")
-    celery_app.worker_main(argv=["worker", "--loglevel=info", "--concurrency=2"])
+    default_queue_name = celery_app.conf.get("task_default_queue", "celery")
+    celery_app.worker_main(
+        argv=[
+            "worker",
+            "--loglevel=info",
+            "--concurrency=2",
+            f"--queues={default_queue_name},{EMAIL_QUEUE_NAME}",
+        ]
+    )
 
 
 if __name__ == "__main__":