+++ tests/ops/api/v1/endpoints/test_connection_config_endpoints.py
@@ -1215,10 +1215,8 @@ class TestPutConnectionConfigSecrets:
             kwargs["action_type"] == EmailActionType.EMAIL_ERASURE_REQUEST_FULFILLMENT
         )
         assert kwargs["to_email"] == "test@example.com"
-        assert kwargs["email_body_params"] == {
-            CollectionAddress(
-                "test_dataset", "test_collection"
-            ): CheckpointActionRequired(
+        assert kwargs["email_body_params"] == [
+            CheckpointActionRequired(
                 step=CurrentStep.erasure,
                 collection=CollectionAddress("test_dataset", "test_collection"),
                 action_needed=[
@@ -1229,4 +1227,4 @@ class TestPutConnectionConfigSecrets:
                     )
                 ],
             )
-        }
+        ]