+++ tests/ops/service/connectors/test_email_connector.py
@@ -9,7 +9,6 @@ from fidesops.ops.graph.config import (
 )
 from fidesops.ops.graph.graph import Edge, Node
 from fidesops.ops.graph.traversal import TraversalNode, artificial_traversal_node
-from fidesops.ops.models.policy import ActionType, Policy, Rule, RuleTarget
 from fidesops.ops.models.privacy_request import ManualAction
 from fidesops.ops.service.connectors import EmailConnector